+++ CHANGELOG.md
@@ -14,12 +14,14 @@ Supported `polkadot-sdk` rev: `274a781e8ca1a9432c7ec87593bd93214abbff50`
 - The `resolc` executable prints the help by default.
 - Removed support for legacy EVM assembly (EVMLA) translation.
 - integration: identify cached code blobs on source code to fix potential confusions.
+- Setting base, include or allow paths in emscripten is now a hard error.
 
 ### Fixed
 - Solidity: Add the solc `--libraries` files to sources.
 - A data race in tests.
 - Fix `broken pipe` errors.
 - llvm-builder: Allow warnings.
+- solidity: Fix the custom compiler warning messages.
 
 ## v0.1.0-dev.9